# backend/app/domains/contracts/router.py
from typing import Annotated
from uuid import UUID

from fastapi import APIRouter, BackgroundTasks, Depends, Response, status
from sqlalchemy.orm import Session

from app.core.config import Settings, get_settings
from app.core.database import get_db
from app.core.errors import ErrorResponse
from app.domains.contracts.schemas import (
    ContractConfirm,
    ContractDetailRead,
    ContractListRead,
    PaymentStatusUpdate,
)
from app.domains.contracts.service import ContractManagementService, ContractQueryService
from app.domains.rag.service import process_contract_index

router = APIRouter(prefix="/contracts", tags=["contracts"])


@router.get("", response_model=ContractListRead)
def list_contracts(
    db: Annotated[Session, Depends(get_db)],
    configuration: Annotated[Settings, Depends(get_settings)],
) -> ContractListRead:
    return ContractQueryService(db, configuration).list_contracts()


@router.get(
    "/{contract_id}",
    response_model=ContractDetailRead,
    responses={
        404: {"model": ErrorResponse, "description": "Contract not found"},
    },
)
def get_contract(
    contract_id: UUID,
    db: Annotated[Session, Depends(get_db)],
    configuration: Annotated[Settings, Depends(get_settings)],
) -> ContractDetailRead:
    return ContractQueryService(db, configuration).get_contract(contract_id)


@router.put(
    "/{contract_id}",
    response_model=ContractDetailRead,
    responses={
        404: {"model": ErrorResponse, "description": "Contract not found"},
        422: {"model": ErrorResponse, "description": "Invalid contract data"},
    },
)
def update_contract(
    contract_id: UUID,
    payload: ContractConfirm,
    background_tasks: BackgroundTasks,
    db: Annotated[Session, Depends(get_db)],
    configuration: Annotated[Settings, Depends(get_settings)],
) -> ContractDetailRead:
    contract = ContractManagementService(db, configuration).update(contract_id, payload)
    background_tasks.add_task(process_contract_index, contract.id, configuration)
    return contract


@router.delete(
    "/{contract_id}",
    status_code=status.HTTP_204_NO_CONTENT,
    responses={404: {"model": ErrorResponse, "description": "Contract not found"}},
)
def delete_contract(
    contract_id: UUID,
    db: Annotated[Session, Depends(get_db)],
    configuration: Annotated[Settings, Depends(get_settings)],
) -> Response:
    ContractManagementService(db, configuration).delete(contract_id)
    return Response(status_code=status.HTTP_204_NO_CONTENT)


@router.patch(
    "/{contract_id}/payments/{payment_id}",
    response_model=ContractDetailRead,
    responses={404: {"model": ErrorResponse, "description": "Contract or payment not found"}},
)
def update_payment_status(
    contract_id: UUID,
    payment_id: UUID,
    payload: PaymentStatusUpdate,
    db: Annotated[Session, Depends(get_db)],
    configuration: Annotated[Settings, Depends(get_settings)],
) -> ContractDetailRead:
    return ContractManagementService(db, configuration).update_payment_status(
        contract_id,
        payment_id,
        payload,
    )